fix(get_thread): ask for the headers a metadata fetch is parsed for
Review of the previous commit. `authentication: { unchecked: true }` is a claim
about the message — nobody checked this mail. On a `full: false` fetch it could
have been produced by the request instead: `format=metadata` returns headers at
the API's discretion, and a response without `Authentication-Results` parses to
exactly the same words. A fact about what we asked for, wearing the sentence
reserved for a fact about the mail.

That is the confusion `full: false` already avoids elsewhere, and it is the one
that nearly archived an invoice as attachment-less: the body and attachment
fields are omitted rather than reported empty, precisely because a request that
never looked must not answer as though it had.

So the metadata fetch now names the headers it is parsed for. PARSED_HEADERS sits
next to parseMessage, and a test parses one message twice — with every header, and
with only the requested ones — requiring the same result, so a header added to the
parse without being added to the list fails the suite rather than going quietly
empty on every metadata fetch.

Also states the assumption the first-report rule rests on: that Gmail returns
payload.headers in message order. It does, and every trace-header convention is
built that way, but it is not a documented guarantee — which is why the result
carries authservId and otherReports rather than a bare verdict.

+++ src/tools.ts
@@ -121,7 +121,7 @@ const authenticationSchema = z.object({
   headerFrom: z.string().optional(),
   returnPath: z.string().optional(),
   alsoReported: z.array(z.string()).optional(),
-  otherReports: z.number().optional(),
+  otherReports: z.number().int().optional(),
   unchecked: z.literal(true).optional(),
 });
